odoo-ingestion/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
faker>=18.0.0

# Odoo integration
xmlrpc>=1.0.0
requests>=2.31.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0

# Type checking
pydantic>=2.0.0